change power field to FloatField in CarListing and Prediction models

+++ backend/predictor/models.py
@@ -12,12 +12,12 @@ class CarListing(models.Model):
     brand = models.CharField(max_length=100)
     car_model = models.CharField(max_length=100)
     year_of_production = models.IntegerField(validators=[MinValueValidator(1900), MaxValueValidator(2100)])
-    mileage = models.IntegerField(validators=[MinValueValidator(0)])
+    mileage = models.IntegerField(validators=[MinValueValidator(0)]) # [km]
     fuel_type = models.CharField(max_length=50)
     transmission = models.CharField(max_length=50)
     body = models.CharField(max_length=50)
-    engine_capacity = models.FloatField(validators=[MinValueValidator(0.0)])
-    power = models.IntegerField(validators=[MinValueValidator(0)])
+    engine_capacity = models.FloatField(validators=[MinValueValidator(0.0)]) # [dm3]
+    power = models.FloatField(validators=[MinValueValidator(0.0)]) # [hp]
     number_of_doors = models.IntegerField(validators=[MinValueValidator(1), MaxValueValidator(10)])
     color = models.CharField(max_length=50)
     price = models.FloatField()
@@ -41,7 +41,7 @@ class Prediction(models.Model):
     transmission = models.CharField(max_length=50)
     body = models.CharField(max_length=50)
     engine_capacity = models.FloatField(validators=[MinValueValidator(0.0)]) # [dm3]
-    power = models.IntegerField(validators=[MinValueValidator(0)]) # [hp]
+    power = models.FloatField(validators=[MinValueValidator(0.0)]) # [hp]
     number_of_doors = models.IntegerField(validators=[MinValueValidator(1), MaxValueValidator(10)])
     color = models.CharField(max_length=50)
     predicted_price = models.FloatField(null=True, blank=True)  # filled after prediction
